refactor: split post content types
