+++ src/components/ui/postList.tsx
@@ -4,7 +4,7 @@ import type { ComponentProps } from 'react';
 import { twMerge } from 'tailwind-merge';
 import { ROUTES } from '@/constants/menu.constants';
 import { POST_CARD_INTERACTION_CLASS } from '@/constants/style.constants';
-import type { Post } from '@/types/content.types';
+import type { Post } from '@/types/post.types';
 import { PostViews } from './postViews';
 import { RelativeTime } from './relativeTime';
