+++ src/types/mdx.d.ts
@@ -1,8 +1,9 @@
 declare module '*.mdx' {
   import type { MDXContent } from 'mdx/types';
-  import type { PostMetadata } from '@/types/content.types';
+  import type { PostMetadata } from '@/types/post.types';
+  import type { ProjectMetadata } from '@/types/project.types';
 
   const content: MDXContent;
   export default content;
-  export const metadata: PostMetadata;
+  export const metadata: PostMetadata | ProjectMetadata;
 }
